+++ charts/fluid/fluid/CHANGELOG.md
@@ -19,3 +19,5 @@
 * Add debug info for csi
 * Make mount root configurable
 * Update HCFS URL
+* Split the controller into alluxio runtime and dataset
+* Implement DataLoad CRD and DataLoad controller